Skip folders called `test`.
Some packages, like Mozilla’s Add-On SDK and its dependees, use this
folder to store test node modules.

Ideally, there would be a way to detect if a module were local only,
and not add its entry to the `package.json`.

This fix also removes a considerable number of resolved but nonetheless
unnecessary packages from the `package.json` for Workpop.