complete rewrite to use `build`
This is a complete rewrite of demeteorizer in an effort to use the
meteor build tool over the built-in bundle command. Functionality
remains roughly the same, but there is still a lot of testing required
for how the CLI hooks up.

The module has been restructured and updated to have 100% test
coverage. CLI functionality has been moved to the demeteorizer bin
over a separate module and each bit of functionality is its own
module.